refactor #1: global 패키지 생성 및 util 패키지 이동
