+++ 1_Week_Mission/multbooks/src/main/java/com/ebook/multbooks/app/member/controller/MemberController.java
@@ -4,8 +4,8 @@ import com.ebook.multbooks.app.member.dto.JoinFormDto;
 import com.ebook.multbooks.app.member.entity.Member;
 import com.ebook.multbooks.app.member.service.MemberService;
 import com.ebook.multbooks.app.security.dto.MemberContext;
-import com.ebook.multbooks.util.Util;
-import com.ebook.multbooks.util.smtp.EmailService;
+import com.ebook.multbooks.global.util.Util;
+import com.ebook.multbooks.global.smtp.EmailService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
