+++ 1_Week_Mission/multbooks/src/main/java/com/ebook/multbooks/app/security/handler/CustomAuthFailureHandler.java
@@ -1,6 +1,6 @@
 package com.ebook.multbooks.app.security.handler;
 
-import com.ebook.multbooks.util.Util;
+import com.ebook.multbooks.global.util.Util;
 import org.springframework.security.authentication.AuthenticationCredentialsNotFoundException;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.InternalAuthenticationServiceException;
