+++ 1_Week_Mission/multbooks/src/main/java/com/ebook/multbooks/global/smtp/EmailService.java
@@ -1,4 +1,4 @@
-package com.ebook.multbooks.util.smtp;
+package com.ebook.multbooks.global.smtp;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.mail.MailSender;
